# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rukhsty - Jordanian Digital Licensing Platform API
servers:
  - url: /api
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      responses:
        "200":
          description: User's applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
  /applications/{id}/submit:
    post:
      operationId: submitApplication
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/select-training-center:
    post:
      operationId: selectTrainingCenter
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectCenterBody"
      responses:
        "200":
          description: Center selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # DOCUMENTS
...
components:
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        code:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, nameAr, nameEn, code, isActive]
    LicenseCategory:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        minimumAge:
          type: integer
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
      required: [id, code, nameAr, nameEn, minimumAge, isActive]

    # APPLICATIONS
    CreateApplicationBody:
      type: object
      properties:
        serviceId:
          type: string
        licenseCategoryId:
          type: string
          nullable: true
        governorate:
          type: string
        residenceArea:
          type: string
      required: [serviceId, governorate, residenceArea]
    Application:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceId:
          type: string
          nullable: true
        licenseCategoryId:
          type: string
          nullable: true
        applicationNumber:
          type: string
        status:
          type: string
        currentStep:
          type: string
        governorate:
          type: string
          nullable: true
        residenceArea:
          type: string
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        submittedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
        licenseCategory:
          $ref: "#/components/schemas/LicenseCategory"
          nullable: true
      required: [id, userId, applicationNumber, status, currentStep, createdAt, updatedAt]
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/ApplicationStep"
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            appointments:
              type: array
              items:
                $ref: "#/components/schemas/Appointment"
            trainingRecord:
              $ref: "#/components/schemas/TrainingRecord"
              nullable: true
            medicalTest:
              $ref: "#/components/schemas/MedicalTest"
              nullable: true
            exams:
              type: array
              items:
                $ref: "#/components/schemas/Exam"
    ApplicationStep:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        stepKey:
          type: string
        stepNameAr:
          type: string
        stepNameEn:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, COMPLETED, FAILED, SKIPPED]
        orderNumber:
          type: integer
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, applicationId, stepKey, stepNameAr, stepNameEn, status, orderNumber]
    SelectCenterBody:
      type: object
      properties:
        centerId:
          type: string
      required: [centerId]

    # DOCUMENTS
...
    Document:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        applicationId:
          type: string
          nullable: true
        documentType:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
          nullable: true
        mimeType:
          type: string
          nullable: true
        verificationStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        rejectionReason:
          type: string
          nullable: true
        verifiedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, documentType, fileUrl, verificationStatus, createdAt]

    # CENTERS
    Center:
      type: object
      properties:
        id:
          type: string
        centerType:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
          nullable: true
        governorate:
          type: string
        city:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, centerType, nameAr, governorate, isActive, createdAt]
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        userId:
          type: string
        centerId:
          type: string
          nullable: true
        appointmentType:
          type: string
        appointmentDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        queueNumber:
          type: integer
          nullable: true
        status:
          type: string
          enum: [BOOKED, CHECKED_IN, COMPLETED, CANCELLED, NO_SHOW, FAILED, PASSED]
        notes:
          type: string
          nullable: true
        center:
          $ref: "#/components/schemas/Center"
          nullable: true
        createdAt:
          type: string
      required: [id, applicationId, userId, appointmentType, appointmentDate, startTime, endTime, status, createdAt]
...
    TrainingRecord:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        centerId:
          type: string
          nullable: true
        theoreticalLessonsRequired:
          type: integer
        practicalLessonsRequired:
          type: integer
        theoreticalLessonsCompleted:
          type: integer
        practicalLessonsCompleted:
          type: integer
        status:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        instructorName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        center:
          $ref: "#/components/schemas/Center"
          nullable: true
      required: [id, applicationId, theoreticalLessonsRequired, practicalLessonsRequired, theoreticalLessonsCompleted, practicalLessonsCompleted, status]
...
    MedicalTest:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        centerId:
          type: string
          nullable: true
        officerId:
          type: string
          nullable: true
        result:
          type: string
          enum: [DOES_NOT_NEED_GLASSES, NEEDS_GLASSES, NOT_FIT_TO_DRIVE, PASS_NO_GLASSES, PASS_WITH_GLASSES, FAILED_NOT_ALLOWED, NEEDS_RECHECK]
        leftEyeScore:
          type: string
          nullable: true
        rightEyeScore:
          type: string
          nullable: true
        requiresGlasses:
          type: boolean
        isAllowedToDrive:
          type: boolean
        notes:
          type: string
          nullable: true
        testedAt:
          type: string
      required: [id, applicationId, result, requiresGlasses, isAllowedToDrive, testedAt]
...
    Exam:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        centerId:
          type: string
          nullable: true
        officerId:
          type: string
          nullable: true
        examType:
          type: string
          enum: [THEORY, PRACTICAL]
        score:
          type: string
          nullable: true
        maxScore:
          type: string
          nullable: true
        result:
          type: string
          enum: [PASSED, FAILED, ABSENT]
        attemptNumber:
          type: integer
        notes:
          type: string
          nullable: true
        examDate:
          type: string
      required: [id, applicationId, examType, result, attemptNumber, examDate]
...